separated posts into separate component
